refactor: KEEP-356 group remaining system steps under lib/workflow/nodes
Moves the remaining lib/steps/ files into per-node directories. Pure
relocation. lib/steps/step-handler.ts is intentionally left in place
since it is shared infrastructure used by plugin steps too; it will
move to lib/workflow/executor/step-handler.ts in the executor batch.

Moves:
- lib/steps/http-request.ts          -> lib/workflow/nodes/http-request/step.ts
- lib/steps/database-query.ts        -> lib/workflow/nodes/database-query/step.ts
- lib/steps/collect.ts               -> lib/workflow/nodes/collect/step.ts
- lib/steps/enrich-explorer-links.ts -> lib/workflow/nodes/enrich-explorer-links/step.ts
- lib/steps/trigger.ts               -> lib/workflow/nodes/trigger/step.ts
- lib/steps/__tests__/collect.test.ts -> lib/workflow/nodes/collect/__tests__/step.test.ts
