+++ app/api/execute/_lib/action-resolver.ts
@@ -22,12 +22,12 @@ const SYSTEM_ACTIONS: Record<
   { importer: () => Promise<any>; stepFunction: string; label: string }
 > = {
   "Database Query": {
-    importer: () => import("@/lib/steps/database-query"),
+    importer: () => import("@/lib/workflow/nodes/database-query/step"),
     stepFunction: "databaseQueryStep",
     label: "Database Query",
   },
   "HTTP Request": {
-    importer: () => import("@/lib/steps/http-request"),
+    importer: () => import("@/lib/workflow/nodes/http-request/step"),
     stepFunction: "httpRequestStep",
     label: "HTTP Request",
   },
